Updated versions - WordPress 6.9.0 + PHP 8.3 + MariaDB 11.8.3 and plugins Akismet 5.6, Blocksy Companion 2.1.23, Elementor 3.34.0, WooCommerce 10.4.3, WPForms Lite 1.9.8.7

+++ wp-content/plugins/woocommerce/includes/class-wc-session-handler.php
@@ -1,4 +1,4 @@
-<?php // phpcs:ignore Generic.PHP.RequireStrictTypes.MissingDeclaration
+<?php
 /**
  * Handle data for the current customers session.
  * Implements the WC_Session abstract class.
@@ -7,10 +7,12 @@
  *
  * @class    WC_Session_Handler
  * @package  WooCommerce\Classes
- * @internal "Missing required strict_types declaration" rule has been ignored to prevent errors with `StringUtil::starts_with` when used on a nonce action which could be -1 rather than a string.
  */
 
+declare(strict_types=1);
+
 use Automattic\Jetpack\Constants;
+use Automattic\WooCommerce\Internal\Features\FeaturesController;
 use Automattic\WooCommerce\Utilities\StringUtil;
 use Automattic\WooCommerce\StoreApi\Utilities\CartTokenUtils;
 
@@ -26,21 +28,21 @@ class WC_Session_Handler extends WC_Session {
 	 *
 	 * @var string cookie name
 	 */
-	protected $_cookie; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
+	protected $_cookie = ''; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
 	 * Stores session expiry.
 	 *
-	 * @var string session due to expire timestamp
+	 * @var int session due to expire timestamp
 	 */
-	protected $_session_expiring; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
+	protected $_session_expiring = 0; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
 	 * Stores session due to expire timestamp.
 	 *
-	 * @var string session expiration timestamp
+	 * @var int session expiration timestamp
 	 */
-	protected $_session_expiration; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
+	protected $_session_expiration = 0; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
 	 * True when the cookie exists.
@@ -54,7 +56,7 @@ class WC_Session_Handler extends WC_Session {
 	 *
 	 * @var string Custom session table name
 	 */
-	protected $_table; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
+	protected $_table = ''; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
 	/**
 	 * Constructor for the session class.
@@ -67,7 +69,7 @@ class WC_Session_Handler extends WC_Session {
 		 *
 		 * @param string $cookie Cookie name.
 		 */
-		$this->_cookie = apply_filters( 'woocommerce_cookie', 'wp_woocommerce_session_' . COOKIEHASH );
+		$this->_cookie = (string) apply_filters( 'woocommerce_cookie', 'wp_woocommerce_session_' . COOKIEHASH );
 		$this->_table  = $GLOBALS['wpdb']->prefix . 'woocommerce_sessions';
 		$this->set_session_expiration();
 	}
@@ -88,6 +90,7 @@ class WC_Session_Handler extends WC_Session {
 	protected function init_hooks() {
 		add_action( 'woocommerce_set_cart_cookies', array( $this, 'set_customer_session_cookie' ), 10 );
 		add_action( 'wp', array( $this, 'maybe_set_customer_session_cookie' ), 99 );
+		add_action( 'template_redirect', array( $this, 'destroy_session_if_empty' ), 999 );
 		add_action( 'shutdown', array( $this, 'save_data' ), 20 );
 		add_action( 'wp_logout', array( $this, 'destroy_session' ) );
 
@@ -116,7 +119,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @return bool
 	 */
 	private function init_session_from_request() {
-		$session_token = wc_clean( wp_unslash( $_GET['session'] ?? '' ) ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		$session_token = is_string( $_GET['session'] ?? '' ) ? wc_clean( wp_unslash( $_GET['session'] ?? '' ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 
 		if ( empty( $session_token ) || ! CartTokenUtils::validate_cart_token( $session_token ) ) {
 			return false;
@@ -137,7 +140,7 @@ class WC_Session_Handler extends WC_Session {
 				return false;
 			}
 
-			$cookie_session_data = $this->get_session( $cookie[0] );
+			$cookie_session_data = (array) $this->get_session( $cookie[0], array() );
 
 			// Cookie session was originally created via this token. Return and use cookie session to prevent creating a new clone.
 			if ( isset( $cookie_session_data['previous_customer_id'] ) && $cookie_session_data['previous_customer_id'] === $payload['user_id'] ) {
@@ -170,34 +173,28 @@ class WC_Session_Handler extends WC_Session {
 
 		// Customer ID will be an MD5 hash id this is a guest session.
 		$this->_customer_id        = $cookie[0];
-		$this->_session_expiration = $cookie[1];
-		$this->_session_expiring   = $cookie[2];
+		$this->_session_expiration = (int) $cookie[1];
+		$this->_session_expiring   = (int) $cookie[2];
 		$this->_has_cookie         = true;
 
 		$this->restore_session_data();
 
 		/**
 		 * This clears the session if the cookie is invalid.
-		 *
-		 * Previously this also cleared the session when $this->_data was empty, and the cart was not yet initialised,
-		 * however this caused a conflict with WooCommerce Payments session handler which overrides this class.
-		 *
-		 * Ref: https://github.com/woocommerce/woocommerce/pull/57652
-		 * See also: https://github.com/woocommerce/woocommerce/pull/59530
 		 */
 		if ( ! $this->is_session_cookie_valid() ) {
 			$this->destroy_session();
 		}
 
 		// If the user logs in, update session.
-		if ( is_user_logged_in() && strval( get_current_user_id() ) !== $this->_customer_id ) {
-			$this->migrate_guest_session_to_user_session( get_current_user_id() );
+		if ( is_user_logged_in() && (string) get_current_user_id() !== $this->get_customer_id() ) {
+			$this->migrate_guest_session_to_user_session();
 		}
 
 		// Update session if its close to expiring.
 		if ( $this->is_session_expiring() ) {
 			$this->set_session_expiration();
-			$this->update_session_timestamp( $this->_customer_id, $this->_session_expiration );
+			$this->update_session_timestamp( $this->get_customer_id(), $this->_session_expiration );
 		}
 	}
 
@@ -207,7 +204,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @param string $clone_from_customer_id The customer ID to clone from.
 	 */
 	private function clone_session_data( string $clone_from_customer_id ) {
-		$session_data                         = $this->get_session( $clone_from_customer_id, array() );
+		$session_data                         = (array) $this->get_session( $clone_from_customer_id, array() );
 		$session_data['previous_customer_id'] = $clone_from_customer_id;
 		$session_data                         = array_diff_key( $session_data, array( 'customer' => true ) );
 		$this->_data                          = $session_data;
@@ -216,7 +213,7 @@ class WC_Session_Handler extends WC_Session {
 	}
 
 	/**
-	 * Migrates a guest session to a user session.
+	 * Migrates a guest session to the current user session.
 	 */
 	private function migrate_guest_session_to_user_session() {
 		$guest_session_id = $this->_customer_id;
@@ -266,7 +263,7 @@ class WC_Session_Handler extends WC_Session {
 		 * @param array $session_data The session data loaded from storage.
 		 * @return array Modified session data to be used for initialization.
 		 */
-		$this->_data = apply_filters( 'woocommerce_restored_session_data', $session_data );
+		$this->_data = (array) apply_filters( 'woocommerce_restored_session_data', $session_data );
 	}
 
 	/**
@@ -281,12 +278,12 @@ class WC_Session_Handler extends WC_Session {
 		}
 
 		// If user has logged out, session cookie is invalid.
-		if ( ! is_user_logged_in() && ! $this->is_customer_guest( $this->_customer_id ) ) {
+		if ( ! is_user_logged_in() && ! $this->is_customer_guest( $this->get_customer_id() ) ) {
 			return false;
 		}
 
 		// Session from a different user is not valid. (Although from a guest user will be valid).
-		if ( is_user_logged_in() && ! $this->is_customer_guest( $this->_customer_id ) && strval( get_current_user_id() ) !== $this->_customer_id ) {
+		if ( is_user_logged_in() && ! $this->is_customer_guest( $this->get_customer_id() ) && (string) get_current_user_id() !== $this->get_customer_id() ) {
 			return false;
 		}
 
@@ -313,7 +310,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @param string $message Value to hash.
 	 * @return string Hashed value.
 	 */
-	private function hash( $message ) {
+	private function hash( string $message ) {
 		if ( function_exists( 'wp_fast_hash' ) ) {
 			return wp_fast_hash( $message );
 		}
@@ -329,7 +326,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @param string $hash Hash to verify.
 	 * @return bool Whether the hash is valid.
 	 */
-	private function verify_hash( $message, $hash ) {
+	private function verify_hash( string $message, string $hash ) {
 		if ( function_exists( 'wp_verify_fast_hash' ) ) {
 			return wp_verify_fast_hash( $message, $hash );
 		}
@@ -347,8 +344,8 @@ class WC_Session_Handler extends WC_Session {
 	 */
 	public function set_customer_session_cookie( $set ) {
 		if ( $set ) {
-			$cookie_hash  = $this->hash( $this->_customer_id . '|' . $this->_session_expiration );
-			$cookie_value = $this->_customer_id . '|' . $this->_session_expiration . '|' . $this->_session_expiring . '|' . $cookie_hash;
+			$cookie_hash  = $this->hash( $this->get_customer_id() . '|' . (string) $this->_session_expiration );
+			$cookie_value = $this->get_customer_id() . '|' . (string) $this->_session_expiration . '|' . (string) $this->_session_expiring . '|' . $cookie_hash;
 
 			if ( ! isset( $_COOKIE[ $this->_cookie ] ) || $_COOKIE[ $this->_cookie ] !== $cookie_value ) {
 				wc_setcookie( $this->_cookie, $cookie_value, $this->_session_expiration, $this->use_secure_cookie(), true );
@@ -372,7 +369,7 @@ class WC_Session_Handler extends WC_Session {
 		 *
 		 * @param bool $use_secure_cookie Whether to use a secure cookie.
 		 */
-		return apply_filters( 'wc_session_use_secure_cookie', wc_site_is_https() && is_ssl() );
+		return (bool) apply_filters( 'wc_session_use_secure_cookie', wc_site_is_https() && is_ssl() );
 	}
 
 	/**
@@ -412,7 +409,6 @@ class WC_Session_Handler extends WC_Session {
 		$expiring_seconds = intval( apply_filters( 'wc_session_expiring', $default_expiring_seconds ) ) ?: $default_expiring_seconds; // phpcs:ignore Universal.Operators.DisallowShortTernary.Found
 
 		if ( $expiring_seconds > $max_expiring_seconds ) {
-			$expiring_seconds       = $max_expiring_seconds;
 			$session_limit_exceeded = true;
 		}
 		/**
@@ -425,14 +421,19 @@ class WC_Session_Handler extends WC_Session {
 
 		// We limit the expiration time to 30 days to avoid performance issues and the session table growing too large.
 		if ( $expiration_seconds > $max_expiration_seconds ) {
-			$expiration_seconds     = $max_expiration_seconds;
 			$session_limit_exceeded = true;
 		}
 
 		if ( $session_limit_exceeded ) {
 			$transient_key = 'wc_session_handler_warning';
 			if ( false === get_transient( $transient_key ) ) {
-				wc_get_logger()->warning( sprintf( 'Keeping sessions for longer than %d days results in performance isues, expiry has been capped.', $max_expiration_seconds / DAY_IN_SECONDS ), array( 'source' => 'wc_session_handler' ) );
+				wc_get_logger()->warning(
+					sprintf(
+						'Keeping sessions for longer than %d days can cause performance issues and larger session tables. Monitor usage and adjust lifetimes via the wc_session_expiring and wc_session_expiration filters as needed.',
+						$max_expiration_seconds / DAY_IN_SECONDS
+					),
+					array( 'source' => 'wc_session_handler' )
+				);
 				set_transient( $transient_key, true, $max_expiration_seconds );
 			}
 		}
@@ -442,8 +443,7 @@ class WC_Session_Handler extends WC_Session {
 			$expiring_seconds = $expiration_seconds * 0.9;
 		}
 
-		$this->_session_expiring = time() + $expiring_seconds;
-
+		$this->_session_expiring   = time() + $expiring_seconds;
 		$this->_session_expiration = time() + $expiration_seconds;
 	}
 
@@ -453,7 +453,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @return string
 	 */
 	public function generate_customer_id() {
-		return is_user_logged_in() ? strval( get_current_user_id() ) : wc_rand_hash( 't_', 30 );
+		return is_user_logged_in() ? (string) get_current_user_id() : wc_rand_hash( 't_', 30 );
 	}
 
 	/**
@@ -476,8 +476,8 @@ class WC_Session_Handler extends WC_Session {
 	public function get_customer_unique_id() {
 		$customer_id = '';
 
-		if ( $this->has_session() && $this->_customer_id ) {
-			$customer_id = $this->_customer_id;
+		if ( $this->has_session() && $this->get_customer_id() ) {
+			$customer_id = $this->get_customer_id();
 		} elseif ( is_user_logged_in() ) {
 			$customer_id = (string) get_current_user_id();
 		}
@@ -531,7 +531,7 @@ class WC_Session_Handler extends WC_Session {
 	 * @return array
 	 */
 	public function get_session_data() {
-		return $this->has_session() ? (array) $this->get_session( $this->_customer_id, array() ) : array();
+		return $this->has_session() ? (array) $this->get_session( $this->get_customer_id(), array() ) : array();
 	}
 
 	/**
@@ -560,12 +560,12 @@ class WC_Session_Handler extends WC_Session {
 					'INSERT INTO %i (`session_key`, `session_value`, `session_expiry`) VALUES (%s, %s, %d)
  					ON DUPLICATE KEY UPDATE `session_value` = VALUES(`session_value`), `session_expiry` = VALUES(`session_expiry`)',
 					$this->_table,
-					$this->_customer_id,
+					$this->get_customer_id(),
 					maybe_serialize( $this->_data ),
 					$this->_session_expiration
 				)
 			);
-			wp_cache_set( $this->get_cache_prefix() . $this->_customer_id, $this->_data, WC_SESSION_CACHE_GROUP, $this->_session_expiration - time() );
+			wp_cache_set( $this->get_cache_prefix() . $this->get_customer_id(), $this->_data, WC_SESSION_CACHE_GROUP, $this->_session_expiration - time() );
 			$this->_dirty = false;
 
 			/**
@@ -573,7 +573,7 @@ class WC_Session_Handler extends WC_Session {
 			 * self::init_session_cookie() upon user login detection initially occurs. However, since some third-party
 			 * extensions override this method, relocating this logic could break backward compatibility.
 			 */
-			if ( ! empty( $old_session_key ) && $this->_customer_id !== $old_session_key && ! is_object( get_user_by( 'id', $old_session_key ) ) ) {
+			if ( ! empty( $old_session_key ) && $this->get_customer_id() !== $old_session_key && ! is_object( get_user_by( 'id', $old_session_key ) ) ) {
 				$this->delete_session( $old_session_key );
 			}
 		}
@@ -583,7 +583,7 @@ class WC_Session_Handler extends WC_Session {
 	 * Destroy all session data.
 	 */
 	public function destroy_session() {
-		$this->delete_session( $this->_customer_id );
+		$this->delete_session( $this->get_customer_id() );
 		$this->forget_session();
 		$this->set_session_expiration();
 	}
@@ -610,13 +610,13 @@ class WC_Session_Handler extends WC_Session {
 	 * This filter runs everything `wp_verify_nonce()` and `wp_create_nonce()` gets called.
 	 *
 	 * @since 5.3.0
-	 * @param int    $uid    User ID.
-	 * @param string $action The nonce action.
+	 * @param int        $uid    User ID.
+	 * @param int|string $action The nonce action.
 	 * @return int|string
 	 */
 	public function maybe_update_nonce_user_logged_out( $uid, $action ) {
-		if ( StringUtil::starts_with( $action, 'woocommerce' ) ) {
-			return $this->has_session() && $this->_customer_id ? $this->_customer_id : $uid;
+		if ( is_string( $action ) && StringUtil::starts_with( $action, 'woocommerce' ) ) {
+			return $this->has_session() && $this->get_customer_id() ? $this->get_customer_id() : $uid;
 		}
 		return $uid;
 	}
@@ -627,9 +627,23 @@ class WC_Session_Handler extends WC_Session {
 	public function cleanup_sessions() {
 		global $wpdb;
 
-		$wpdb->query( $wpdb->prepare( 'DELETE FROM %i WHERE session_expiry < %d', $this->_table, time() ) );
+		// Batch size of 100 and sleep time of 10ms = max 100 SQL queries and 10K entries deletion per second.
+		$batch_size            = 100;
+		$deleted_entries_total = 0;
+		do {
+			$deleted_entries_count  = (int) $wpdb->query(
+				$wpdb->prepare(
+					'DELETE FROM %i WHERE session_expiry < %d ORDER BY session_expiry LIMIT %d',
+					$this->_table,
+					time(),
+					$batch_size
+				)
+			);
+			$deleted_entries_total += $deleted_entries_count;
+			usleep( ( 10_000 / $batch_size ) * $deleted_entries_count );
+		} while ( $deleted_entries_count === $batch_size );
 
-		if ( class_exists( 'WC_Cache_Helper' ) ) {
+		if ( $deleted_entries_total > 0 && class_exists( 'WC_Cache_Helper' ) ) {
 			WC_Cache_Helper::invalidate_cache_group( WC_SESSION_CACHE_GROUP );
 		}
 	}
@@ -639,13 +653,13 @@ class WC_Session_Handler extends WC_Session {
 	 *
 	 * @param string $customer_id Customer ID.
 	 * @param mixed  $default_value Default session value.
-	 * @return string|array
+	 * @return mixed Returns either the session data or the default value. Returns false if WP setup is in progress.
 	 */
 	public function get_session( $customer_id, $default_value = false ) {
 		global $wpdb;
 
 		if ( Constants::is_defined( 'WP_SETUP_CONFIG' ) ) {
-			return false;
+			return $default_value;
 		}
 
 		// Try to get it from the cache, it will return false if not present or if object cache not in use.
@@ -670,9 +684,12 @@ class WC_Session_Handler extends WC_Session {
 	/**
 	 * Delete the session from the cache and database.
 	 *
-	 * @param int $customer_id Customer ID.
+	 * @param string $customer_id Customer session ID.
 	 */
 	public function delete_session( $customer_id ) {
+		if ( ! $customer_id ) {
+			return;
+		}
 		$GLOBALS['wpdb']->delete( $this->_table, array( 'session_key' => $customer_id ) );
 		wp_cache_delete( $this->get_cache_prefix() . $customer_id, WC_SESSION_CACHE_GROUP );
 	}
@@ -684,7 +701,51 @@ class WC_Session_Handler extends WC_Session {
 	 * @param int    $timestamp Timestamp to expire the cookie.
 	 */
 	public function update_session_timestamp( $customer_id, $timestamp ) {
+		if ( ! $customer_id ) {
+			return;
+		}
 		$GLOBALS['wpdb']->update( $this->_table, array( 'session_expiry' => $timestamp ), array( 'session_key' => $customer_id ), array( '%d' ) );
+	}
+
+	/**
+	 * Destroys the WooCommerce session if it contains no data for non-logged-in users.
+	 *
+	 * This method helps improve caching performance by removing session cookies when they
+	 * are no longer needed, allowing non-logged-in customers to receive cached pages.
+	 * Only runs if the destroy-empty-sessions feature is enabled.
+	 *
+	 * @return void
+	 *
+	 * @since 10.3.0
+	 *
+	 * @internal For exclusive usage of WooCommerce core, backwards compatibility not guaranteed.
+	 */
+	public function destroy_session_if_empty() {
+		if ( is_user_logged_in() || ! $this->_has_cookie ) {
+			return;
+		}
+
+		if ( ! isset( $_COOKIE[ $this->_cookie ] ) ) {
+			// If $_COOKIE isn't set, then something triggered setting the cookie during this request. So we won't
+			// yet destroy the session if it is empty to expand compatibility at the cost of one additional request being uncached.
+			return;
+		}
+
+		if ( ! empty( $this->_data ) ) {
+			return;
+		}
+
+		if ( is_object( WC()->cart ) && ! WC()->cart->is_empty() ) {
+			// There is a pending cart to save that isn't yet in the session data.
+			return;
+		}
+
+		$feature_controller = wc_get_container()->get( FeaturesController::class );
+		if ( ! $feature_controller->feature_is_enabled( 'destroy-empty-sessions' ) ) {
+			return;
+		}
+
+		$this->destroy_session();
 	}
 
 	/**
@@ -694,6 +755,6 @@ class WC_Session_Handler extends WC_Session {
 	 * @return bool
 	 */
 	private function session_exists( $customer_id ) {
-		return null !== $GLOBALS['wpdb']->get_var( $GLOBALS['wpdb']->prepare( 'SELECT session_key FROM %i WHERE session_key = %s', $this->_table, $customer_id ) );
+		return $customer_id && null !== $GLOBALS['wpdb']->get_var( $GLOBALS['wpdb']->prepare( 'SELECT session_key FROM %i WHERE session_key = %s', $this->_table, $customer_id ) );
 	}
 }
